refactor() rename, move and delete modules
