+++ src/ais_analyzer/__main__.py
@@ -1,8 +1,8 @@
-import my_ais_cli as cli
-import my_ais_handler as ais_handler
-from src.ais_analyzer.handlers import input_handler, output_handler
 
-from commands import statistics
+from src.ais_analyzer.handlers.output_handler import OutputHandler
+from src.ais_analyzer.handlers.input_handler import InputHandler
+from src.ais_analyzer.handlers.cli_handler import MyAISCLI
+from src.ais_analyzer.commands import statistics
 
 
 def __main__():
@@ -17,29 +17,24 @@ def __main__():
 
     # 1. Getting the user arguments from the command line
     # Instantiating the CLI and getting arguments
-    mycli = cli.MyAISCLI()
+    mycli = MyAISCLI()
     args = mycli.get_args(asdict=True)
 
-    # 2. Loading the data using ais handler
-    # Instantiating an AIS handler and passing the CLI argument
-    myais = ais_handler.MyAISHandler()
-    myais.eval_args(args)
-
-    # 2.* Loading the data using input handler
-    my_data = input_handler.InputHandler(args['path'])
+    # 2. Loading the data using input handler
+    my_data = InputHandler(args['path'])
 
     # 3. Calling the command on my data
     # TODODO: Proper input validation
     commands = {'statistics': statistics}
     command = args['commmand']
     params = args['full']
-
     if command in commands.keys():
         # Statistics is the only command, so we call the statistics method on the data
-        res = commands[command](my_data, params)
+        transformed_data = commands[command](my_data, params)
 
     # 4. Output the transformed data
-    output_handler.OutputHandler(args['output'])
+    output_path = args['output']
+    OutputHandler(output_path, transformed_data)
 
 
 if __name__ == "__main__":
